feat(render.tsx): move to lib so we dont have utils and lib folder
